SAK-52658 SiteStats split view service responsibilities
