+++ sitestats/sitestats-impl/src/webapp/WEB-INF/components.xml
@@ -237,18 +237,32 @@
 	<bean id="org.sakaiproject.sitestats.api.view.SiteStatsReportPreviewService"
 		  class="org.sakaiproject.sitestats.impl.view.SiteStatsReportPreviewServiceImpl"/>
 
+	<!-- SiteStatsReportViewMapper _________________________________________________________________ -->
+	<bean id="org.sakaiproject.sitestats.impl.view.SiteStatsReportViewMapper"
+		  class="org.sakaiproject.sitestats.impl.view.SiteStatsReportViewMapper">
+		<property name="reportManager" ref="org.sakaiproject.sitestats.api.report.ReportManager"/>
+		<property name="siteStatsTableMapper" ref="org.sakaiproject.sitestats.impl.view.SiteStatsTableMapper"/>
+	</bean>
+
+	<!-- SiteStatsWidgetCatalog ____________________________________________________________________ -->
+	<bean id="org.sakaiproject.sitestats.impl.view.SiteStatsWidgetCatalog"
+		  class="org.sakaiproject.sitestats.impl.view.SiteStatsWidgetCatalog">
+		<property name="statsManager" ref="org.sakaiproject.sitestats.api.StatsManager"/>
+		<property name="siteStatsToolEventsService" ref="org.sakaiproject.sitestats.api.event.SiteStatsToolEventsService"/>
+		<property name="eventRegistryService" ref="org.sakaiproject.sitestats.api.event.EventRegistryService"/>
+		<property name="siteService" ref="org.sakaiproject.site.api.SiteService"/>
+	</bean>
+
 	<!-- SiteStatsViewService ______________________________________________________________________ -->
 	<bean id="org.sakaiproject.sitestats.api.view.SiteStatsViewService"
 		  class="org.sakaiproject.sitestats.impl.view.SiteStatsViewServiceImpl">
 		<property name="statsAuthz" ref="org.sakaiproject.sitestats.api.StatsAuthz"/>
 		<property name="statsManager" ref="org.sakaiproject.sitestats.api.StatsManager"/>
 		<property name="reportManager" ref="org.sakaiproject.sitestats.api.report.ReportManager"/>
-		<property name="siteStatsToolEventsService" ref="org.sakaiproject.sitestats.api.event.SiteStatsToolEventsService"/>
-		<property name="siteStatsTableMapper" ref="org.sakaiproject.sitestats.impl.view.SiteStatsTableMapper"/>
 		<property name="siteStatsReportPreviewService" ref="org.sakaiproject.sitestats.api.view.SiteStatsReportPreviewService"/>
-		<property name="eventRegistryService" ref="org.sakaiproject.sitestats.api.event.EventRegistryService"/>
-		<property name="siteService" ref="org.sakaiproject.site.api.SiteService"/>
 		<property name="sessionManager" ref="org.sakaiproject.tool.api.SessionManager"/>
+		<property name="siteStatsWidgetCatalog" ref="org.sakaiproject.sitestats.impl.view.SiteStatsWidgetCatalog"/>
+		<property name="siteStatsReportViewMapper" ref="org.sakaiproject.sitestats.impl.view.SiteStatsReportViewMapper"/>
 	</bean>
